Make Texture clearing before IGM optional. (not applying to Rice 6.11 as this will be removed in next commit.)

+++ trunk/B6.0/Plugins/src/Video5.60/RiceVideo.cpp
@@ -32,6 +32,7 @@ extern bool g_bUseSetTextureMem;
 extern DWORD g_maxTextureMemUsage;
 
 #include "../../../config.h"
+extern int loadbForceClearTextures();
 extern bool bloadstate[MAX_SAVE_STATES];
 extern bool bsavestate[MAX_SAVE_STATES];
 extern bool bload1964state[MAX_SAVE_STATES];
@@ -719,8 +720,11 @@ EXPORT void CALL _VIDEO_RICE_560_UpdateScreen (void)
 			if (bloadstate[i]) // This will never be true, unless UHLE accepts plugins. 
 			{
 				try{
-					bPurgeOldBeforeIGM = TRUE;
-					gTextureManager.PurgeOldTextures();
+					if(loadbForceClearTextures())
+					{
+						bPurgeOldBeforeIGM = TRUE;
+						gTextureManager.PurgeOldTextures();
+					}
 					__EMU_Load1964State(i);
 				}catch(...){};
 
@@ -732,8 +736,11 @@ EXPORT void CALL _VIDEO_RICE_560_UpdateScreen (void)
 			else if (bload1964state[i]) 
 			{
 				try{
-					bPurgeOldBeforeIGM = TRUE;
-					gTextureManager.PurgeOldTextures();
+					if(loadbForceClearTextures())
+					{
+						bPurgeOldBeforeIGM = TRUE;
+						gTextureManager.PurgeOldTextures();
+					}
 					__EMU_Load1964State(i);
 				}catch(...){};
 
@@ -745,8 +752,11 @@ EXPORT void CALL _VIDEO_RICE_560_UpdateScreen (void)
 			else if (bloadPJ64state[i]) 
 			{
 				try{
-					bPurgeOldBeforeIGM = TRUE;
-					gTextureManager.PurgeOldTextures();
+					if(loadbForceClearTextures())
+					{
+						bPurgeOldBeforeIGM = TRUE;
+						gTextureManager.PurgeOldTextures();
+					}
 					__EMU_Load1964State(i);
 				}catch(...){};
